Wall Placement and Pathfinding
Pressing right click will allow you to place a wall and the enemies will
find a path around them. They are free to place at the moment and the
price can be changed from the DEFAULT_COST constant in the wall class.
There is no handling for events where the path to the tower is
completely obstructed